+++ src/slow/time_integration/update.py
@@ -4,9 +4,6 @@
 # Author: Y.Takahashi, Hokkaido University
 # Date; 2022/03/31
 
-import numpy as np
-from slow.orbital.orbital import orbital
-
 def update_solution(config, geom_dict, var_conserv, var_dq):
 
   # Main routine